requirements.txt
Django==4.2.7
djongo==1.3.6
pymongo==4.5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8